Further refined implementation of Pokémon evolution
Reworked `PokemonSpecies.evolutions` to work as a computed property, rather than assigning an array of `PokemonEvolution` objects to them.
This implementation would in theory use less memory initially, and more importantly, users can obtain an evolution chain through this.
Previous implementation would have an evolution of an evolution with nil for this property, as its evolutions had not been retrieved, due to value type semantics.

Added implementation of further evolution conditions.

Removed tradeWithItem case on EvolutionConditions, as this is just a combination of trading and having an item, which already exist as two separate cases.

Made more types implement Codable where necessary.